requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
huggingface-hub
python-dotenv
pypdf